move dists
