+++ _contrib/_plot_path.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from sktime.distances._distance import distance_alignment_path, pairwise_distance
+from sktime.dists_kernels._numba_distances._distance import distance_alignment_path, pairwise_distance
 from sktime.utils.dependencies import _check_soft_dependencies
 
 _check_soft_dependencies("matplotlib", severity="warning")
