+++ sktime/clustering/k_medoids.py
@@ -8,7 +8,7 @@ from numpy.random import RandomState
 
 from sktime.clustering.metrics.medoids import medoids
 from sktime.clustering.partitioning import BaseTimeSeriesLloyds
-from sktime.distances import pairwise_distance
+from sktime.dists_kernels._numba_distances import pairwise_distance
 
 
 class TimeSeriesKMedoids(BaseTimeSeriesLloyds):
