+++ sktime/clustering/partitioning/_lloyds.py
@@ -9,8 +9,8 @@ from sklearn.utils.extmath import stable_cumsum
 
 from sktime.clustering.base import BaseClusterer
 from sktime.clustering.metrics.averaging import mean_average
-from sktime.distances import distance_factory, pairwise_distance
-from sktime.distances._ddtw import average_of_slope_transform
+from sktime.dists_kernels._numba_distances import distance_factory, pairwise_distance
+from sktime.dists_kernels._numba_distances._ddtw import average_of_slope_transform
 
 
 def _forgy_center_initializer(
