+++ sktime/dists_kernels/_numba_distances/_erp.py
@@ -4,28 +4,15 @@ from typing import Any
 
 import numpy as np
 
-from sktime.distances.base import (
+from sktime.dists_kernels._numba_distances.base import (
     DistanceAlignmentPathCallable,
     DistanceCallable,
     NumbaDistance,
 )
 
 
-class _EdrDistance(NumbaDistance):
-    """Edit distance for real sequences (EDR) between two time series.
-
-    ERP was adapted in [1] specifically for distances between trajectories. Like LCSS,
-    EDR uses a distance threshold to define when two elements of a series match.
-    However, rather than simply count matches and look for the longest sequence,
-    ERP applies a (constant) penalty for non-matching elements
-    where gaps are inserted to create an optimal alignment.
-
-    References
-    ----------
-    .. [1] Chen L, Ozsu MT, Oria V: Robust and fast similarity search for moving
-    object trajectories. In: Proceedings of the ACM SIGMOD International Conference
-    on Management of Data, 2005
-    """
+class _ErpDistance(NumbaDistance):
+    """Edit distance with real penalty (erp) between two time series."""
 
     def _distance_alignment_path_factory(
         self,
@@ -35,11 +22,12 @@ class _EdrDistance(NumbaDistance):
         window: float = None,
         itakura_max_slope: float = None,
         bounding_matrix: np.ndarray = None,
-        epsilon: float = None,
+        g: float = 0.0,
         **kwargs: Any,
     ) -> DistanceAlignmentPathCallable:
-        """Create a no_python compiled edr alignment path distance callable.
+        """Create a no_python compiled erp distance alignment path callable.
 
+        Similar to LCSS with a different penalty.
         Series should be shape (d, m), where d is the number of dimensions, m the series
         length. Series can be different lengths.
 
@@ -62,70 +50,57 @@ class _EdrDistance(NumbaDistance):
             are ignored. The matrix should be structure so that indexes considered in
             bound should be the value 0. and indexes outside the bounding matrix should
             be infinity.
-        epsilon : float, defaults = None
-            Matching threshold to determine if two subsequences are considered close
-            enough to be considered 'common'. If not specified as per the original paper
-            epsilon is set to a quarter of the maximum standard deviation.
+        g: float, defaults = 0.
+            The reference value to penalise gaps.
         kwargs: Any
             Extra kwargs.
 
         Returns
         -------
         Callable[[np.ndarray, np.ndarray], tuple[np.ndarray, float]]
-            No_python compiled edr distance path callable.
+            No_python compiled wdtw distance path callable.
 
         Raises
         ------
         ValueError
-            If the input time series are not numpy array.
-            If the input time series do not have exactly 2 dimensions.
+            If the input times eries is not a numpy array.
+            If the input time series doesn't have exactly 2 dimensions.
             If the sakoe_chiba_window_radius is not an integer.
             If the itakura_max_slope is not a float or int.
-            If epsilon is not a float.
+            If g is not a float.
         """
-        from sktime.distances._distance_alignment_paths import compute_min_return_path
-        from sktime.distances._edr_numba import _edr_cost_matrix
-        from sktime.distances.lower_bounding import resolve_bounding_matrix
+        from sktime.dists_kernels._numba_distances._distance_alignment_paths import compute_min_return_path
+        from sktime.dists_kernels._numba_distances._erp_numba import _erp_cost_matrix
+        from sktime.dists_kernels._numba_distances.lower_bounding import resolve_bounding_matrix
         from sktime.utils.numba.njit import njit
 
         _bounding_matrix = resolve_bounding_matrix(
             x, y, window, itakura_max_slope, bounding_matrix
         )
-
-        if epsilon is not None and not isinstance(epsilon, float):
-            raise ValueError("The value of epsilon must be a float.")
+        if not isinstance(g, float):
+            raise ValueError("The value of g must be a float.")
 
         if return_cost_matrix is True:
 
             @njit(cache=True)
-            def numba_edr_distance_alignment_path(
+            def numba_erp_distance_alignment_path(
                 _x: np.ndarray, _y: np.ndarray
             ) -> tuple[list, float, np.ndarray]:
-                if epsilon is None:
-                    _epsilon = max(np.std(_x), np.std(_y)) / 4
-                else:
-                    _epsilon = epsilon
-                cost_matrix = _edr_cost_matrix(_x, _y, _bounding_matrix, _epsilon)
+                cost_matrix = _erp_cost_matrix(_x, _y, _bounding_matrix, g)
                 path = compute_min_return_path(cost_matrix, _bounding_matrix)
-                distance = float(cost_matrix[-1, -1] / max(_x.shape[1], _y.shape[1]))
-                return path, distance, cost_matrix
+                return path, cost_matrix[-1, -1], cost_matrix
 
         else:
 
             @njit(cache=True)
-            def numba_edr_distance_alignment_path(
+            def numba_erp_distance_alignment_path(
                 _x: np.ndarray, _y: np.ndarray
             ) -> tuple[list, float]:
-                if epsilon is None:
-                    _epsilon = max(np.std(_x), np.std(_y)) / 4
-                else:
-                    _epsilon = epsilon
-                cost_matrix = _edr_cost_matrix(_x, _y, _bounding_matrix, _epsilon)
+                cost_matrix = _erp_cost_matrix(_x, _y, _bounding_matrix, g)
                 path = compute_min_return_path(cost_matrix, _bounding_matrix)
-                distance = float(cost_matrix[-1, -1] / max(_x.shape[1], _y.shape[1]))
-                return path, distance
+                return path, cost_matrix[-1, -1]
 
-        return numba_edr_distance_alignment_path
+        return numba_erp_distance_alignment_path
 
     def _distance_factory(
         self,
@@ -134,11 +109,12 @@ class _EdrDistance(NumbaDistance):
         window: float = None,
         itakura_max_slope: float = None,
         bounding_matrix: np.ndarray = None,
-        epsilon: float = None,
+        g: float = 0.0,
         **kwargs: Any,
     ) -> DistanceCallable:
-        """Create a no_python compiled edr distance callable.
+        """Create a no_python compiled erp distance callable.
 
+        Similar to LCSS with a different penalty.
         Series should be shape (d, m), where d is the number of dimensions, m the series
         length. Series can be different lengths.
 
@@ -159,47 +135,39 @@ class _EdrDistance(NumbaDistance):
             are ignored. The matrix should be structure so that indexes considered in
             bound should be the value 0. and indexes outside the bounding matrix should
             be infinity.
-        epsilon : float, defaults = None
-            Matching threshold to determine if two subsequences are considered close
-            enough to be considered 'common'. If not specified as per the original paper
-            epsilon is set to a quarter of the maximum standard deviation.
+        g: float, defaults = 0.
+            The reference value to penalise gaps.
         kwargs: Any
             Extra kwargs.
 
         Returns
         -------
         Callable[[np.ndarray, np.ndarray], float]
-            No_python compiled edr distance callable.
+            No_python compiled erp distance callable.
 
         Raises
         ------
         ValueError
-            If the input time series are not numpy array.
-            If the input time series do not have exactly 2 dimensions.
+            If the input time series is not a numpy array.
+            If the input time series doesn't have exactly 2 dimensions.
             If the sakoe_chiba_window_radius is not an integer.
             If the itakura_max_slope is not a float or int.
-            If epsilon is not a float.
+            If g is not a float.
         """
-        from sktime.distances._edr_numba import _edr_cost_matrix
-        from sktime.distances.lower_bounding import resolve_bounding_matrix
+        from sktime.dists_kernels._numba_distances._erp_numba import _erp_cost_matrix
+        from sktime.dists_kernels._numba_distances.lower_bounding import resolve_bounding_matrix
         from sktime.utils.numba.njit import njit
 
         _bounding_matrix = resolve_bounding_matrix(
             x, y, window, itakura_max_slope, bounding_matrix
         )
-
-        if epsilon is not None and not isinstance(epsilon, float):
-            raise ValueError("The value of epsilon must be a float.")
+        if not isinstance(g, float):
+            raise ValueError("The value of g must be a float.")
 
         @njit(cache=True)
-        def numba_edr_distance(_x: np.ndarray, _y: np.ndarray) -> float:
-            if np.array_equal(_x, _y):
-                return 0.0
-            if epsilon is None:
-                _epsilon = max(np.std(_x), np.std(_y)) / 4
-            else:
-                _epsilon = epsilon
-            cost_matrix = _edr_cost_matrix(_x, _y, _bounding_matrix, _epsilon)
-            return float(cost_matrix[-1, -1] / max(_x.shape[1], _y.shape[1]))
+        def numba_erp_distance(_x: np.ndarray, _y: np.ndarray) -> float:
+            cost_matrix = _erp_cost_matrix(_x, _y, _bounding_matrix, g)
 
-        return numba_edr_distance
+            return cost_matrix[-1, -1]
+
+        return numba_erp_distance
