+++ sktime/dists_kernels/_numba_distances/_resolve_metric.py
@@ -4,7 +4,7 @@ from collections.abc import Callable
 
 import numpy as np
 
-from sktime.distances.base import DistanceCallable, MetricInfo, NumbaDistance
+from sktime.dists_kernels._numba_distances.base import DistanceCallable, MetricInfo, NumbaDistance
 
 
 def _resolve_dist_instance(
@@ -185,7 +185,7 @@ def _is_distance_factory_callable(metric: Callable) -> bool:
         Boolean that is true if callable is a valid distance factory and false
         if the callable is an invalid distance factory.
     """
-    from sktime.distances._numba_utils import is_no_python_compiled_callable
+    from sktime.dists_kernels._numba_distances._numba_utils import is_no_python_compiled_callable
 
     is_no_python_compiled = is_no_python_compiled_callable(metric)
     if is_no_python_compiled:
@@ -214,7 +214,7 @@ def _is_no_python_distance_callable(metric: Callable) -> bool:
         Boolean that is true if callable is a valid no_python compiled distance and
         false if the callable is an invalid no_python callable.
     """
-    from sktime.distances._numba_utils import is_no_python_compiled_callable
+    from sktime.dists_kernels._numba_distances._numba_utils import is_no_python_compiled_callable
 
     is_no_python_compiled = is_no_python_compiled_callable(metric)
     if not is_no_python_compiled:
