+++ sktime/dists_kernels/_numba_distances/_squared_numba.py
