+++ sktime/dists_kernels/_numba_distances/_wddtw.py
@@ -4,7 +4,7 @@ from typing import Any
 
 import numpy as np
 
-from sktime.distances.base import (
+from sktime.dists_kernels._numba_distances.base import (
     DistanceAlignmentPathCallable,
     DistanceCallable,
     NumbaDistance,
@@ -80,14 +80,14 @@ class _WddtwDistance(NumbaDistance):
             If the compute derivative callable is not no_python compiled.
             If the value of g is not a float
         """
-        from sktime.distances._distance_alignment_paths import compute_min_return_path
-        from sktime.distances._numba_utils import is_no_python_compiled_callable
-        from sktime.distances._wdtw_numba import _weighted_cost_matrix
-        from sktime.distances.lower_bounding import resolve_bounding_matrix
+        from sktime.dists_kernels._numba_distances._distance_alignment_paths import compute_min_return_path
+        from sktime.dists_kernels._numba_distances._numba_utils import is_no_python_compiled_callable
+        from sktime.dists_kernels._numba_distances._wdtw_numba import _weighted_cost_matrix
+        from sktime.dists_kernels._numba_distances.lower_bounding import resolve_bounding_matrix
         from sktime.utils.numba.njit import njit
 
         if compute_derivative is None:
-            from sktime.distances._ddtw_numba import average_of_slope
+            from sktime.dists_kernels._numba_distances._ddtw_numba import average_of_slope
 
             compute_derivative = average_of_slope
 
@@ -194,13 +194,13 @@ class _WddtwDistance(NumbaDistance):
             If the compute derivative callable is not no_python compiled.
             If the value of g is not a float
         """
-        from sktime.distances._numba_utils import is_no_python_compiled_callable
-        from sktime.distances._wdtw_numba import _weighted_cost_matrix
-        from sktime.distances.lower_bounding import resolve_bounding_matrix
+        from sktime.dists_kernels._numba_distances._numba_utils import is_no_python_compiled_callable
+        from sktime.dists_kernels._numba_distances._wdtw_numba import _weighted_cost_matrix
+        from sktime.dists_kernels._numba_distances.lower_bounding import resolve_bounding_matrix
         from sktime.utils.numba.njit import njit
 
         if compute_derivative is None:
-            from sktime.distances._ddtw_numba import average_of_slope
+            from sktime.dists_kernels._numba_distances._ddtw_numba import average_of_slope
 
             compute_derivative = average_of_slope
