+++ sktime/dists_kernels/_numba_distances/tests/_shared_tests.py
@@ -3,8 +3,8 @@ from collections.abc import Callable
 import numpy as np
 import pytest
 
-from sktime.distances.base import DistanceCallable, NumbaDistance
-from sktime.distances.tests._utils import create_test_distance_numpy
+from sktime.dists_kernels._numba_distances.base import DistanceCallable, NumbaDistance
+from sktime.dists_kernels._numba_distances.tests._utils import create_test_distance_numpy
 from sktime.utils.numba.njit import njit
