+++ sktime/dists_kernels/_numba_distances/tests/test_distance_alignment_path.py
@@ -3,8 +3,8 @@
 import numpy as np
 import pytest
 
-from sktime.distances._distance import _METRIC_INFOS, distance, distance_alignment_path
-from sktime.distances.tests._utils import create_test_distance_numpy
+from sktime.dists_kernels._numba_distances._distance import _METRIC_INFOS, distance, distance_alignment_path
+from sktime.dists_kernels._numba_distances.tests._utils import create_test_distance_numpy
 from sktime.tests.test_switch import run_test_module_changed
 from sktime.utils.dependencies import _check_soft_dependencies
