+++ sktime/dists_kernels/_numba_distances/tests/test_distance_correctness.py
@@ -10,7 +10,7 @@ import pytest
 from numpy.testing import assert_almost_equal
 
 from sktime.datasets import load_basic_motions, load_unit_test
-from sktime.distances._distance import _METRIC_INFOS
+from sktime.dists_kernels._numba_distances._distance import _METRIC_INFOS
 from sktime.tests.test_switch import run_test_for_class, run_test_module_changed
 from sktime.utils.dependencies import _check_soft_dependencies
