+++ sktime/dists_kernels/_numba_distances/tests/test_pairwise_distances.py
@@ -7,14 +7,14 @@ from collections.abc import Callable
 import numpy as np
 import pytest
 
-from sktime.distances._distance import _METRIC_INFOS, pairwise_distance
-from sktime.distances._numba_utils import _make_3d_series
-from sktime.distances.base import MetricInfo, NumbaDistance
-from sktime.distances.tests._shared_tests import (
+from sktime.dists_kernels._numba_distances._distance import _METRIC_INFOS, pairwise_distance
+from sktime.dists_kernels._numba_distances._numba_utils import _make_3d_series
+from sktime.dists_kernels._numba_distances.base import MetricInfo, NumbaDistance
+from sktime.dists_kernels._numba_distances.tests._shared_tests import (
     _test_incorrect_parameters,
     _test_metric_parameters,
 )
-from sktime.distances.tests._utils import create_test_distance_numpy
+from sktime.dists_kernels._numba_distances.tests._utils import create_test_distance_numpy
 from sktime.tests.test_switch import run_test_module_changed
 from sktime.utils.dependencies import _check_soft_dependencies
