+++ sktime/dists_kernels/edit_dist.py
@@ -5,7 +5,7 @@ __author__ = ["fkiraly"]
 
 import numpy as np
 
-from sktime.distances import pairwise_distance
+from sktime.dists_kernels._numba_distances import pairwise_distance
 from sktime.dists_kernels.base import BasePairwiseTransformerPanel
